+++ CHANGELOG.md
@@ -10,4 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 <small>[Compare with latest](https://github.com/ydethe/ragwebui/compare/83348cc6d894474833d75a88988a7a61b109502d...HEAD)</small>
 
+### Added
+
+- Added API keys and remove of deprecated method QdrantClient.search ([7a56da9](https://github.com/ydethe/ragwebui/commit/7a56da902ec471077d7c8d5dd4ab53b000160de6) by Yann de The).
+
 <!-- insertion marker -->